requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
httpx
python-dotenv
email-validator
